# openapi.yaml
openapi: 3.1.0
info:
  title: Permaculture Planner API
  version: 1.1.0
  description: API for raised-bed permaculture garden planning
  contact:
    email: api@permaculture-planner.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.permaculture-planner.com/v1
    description: Production
  - url: https://staging-api.permaculture-planner.com/v1
    description: Staging
  - url: http://localhost:3000/api
    description: Development
security:
  - BearerAuth: []
paths:
  /plans:
    post:
      summary: Create a new garden plan
      operationId: createPlan
      tags: [Plans]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanInput'
            example:
              location:
                city: "Portland, OR"
                usda_zone: "8b"
                last_frost: "2024-04-15"
                first_frost: "2024-11-15"
              area:
                total_sqft: 400
                usable_fraction: 0.8
                shape: "rectangular"
              surface:
                type: "soil"
                sun_hours: 8
                slope: 2
                accessibility_needs: false
              water:
                source: "spigot"
                drip_allowed: true
                sip_interest: false
              crops:
                focus: ["tomatoes", "lettuce", "herbs"]
                time_weekly_minutes: 120
      responses:
        '201':
          description: Plan created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
          headers:
            X-Idempotency-Key:
              schema:
                type: string
              description: Idempotency key for retry safety
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'

    get:
      summary: List user's plans
      operationId: listPlans
      tags: [Plans]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, archived]
      responses:
        '200':
          description: List of plans
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PlanSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PlanInput:
      type: object
      required: [location, area, surface, water, crops]
      properties:
        location:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
            city:
              type: string
            usda_zone:
              type: string
            last_frost:
              type: string
              format: date
            first_frost:
              type: string
              format: date
        area:
          type: object
          required: [total_sqft, usable_fraction, shape]
          properties:
            total_sqft:
              type: number
              minimum: 10
              maximum: 50000
            usable_fraction:
              type: number
              minimum: 0.1
              maximum: 1.0
            shape:
              type: string
              enum: [rectangular, L-shaped, scattered]
        surface:
          type: object
          required: [type, sun_hours, slope]
          properties:
            type:
              type: string
              enum: [soil, hard]
            sun_hours:
              type: number
              minimum: 0
              maximum: 14
            slope:
              type: number
              minimum: 0
              maximum: 45
            accessibility_needs:
              type: boolean
        water:
          type: object
          required: [source]
          properties:
            source:
              type: string
              enum: [spigot, rain, none]
            drip_allowed:
              type: boolean
            sip_interest:
              type: boolean
        crops:
          type: object
          required: [focus, time_weekly_minutes]
          properties:
            focus:
              type: array
              items:
                type: string
            avoid_families:
              type: array
              items:
                type: string
            time_weekly_minutes:
              type: integer
              minimum: 15
              maximum: 600
        materials:
          type: object
          properties:
            lumber_type:
              type: string
              enum: [cedar, pine, treated]
            budget_tier:
              type: string
              enum: [budget, standard, premium]
    PlanResponse:
      type: object
      properties:
        planId:
          type: string
          format: uuid
        success:
          type: boolean
        summary:
          type: object
          properties:
            beds:
              type: integer
            totalArea:
              type: number
            estimatedCost:
              type: object
              properties:
                low:
                  type: number
                high:
                  type: number
            warnings:
              type: array
              items:
                type: string
            suggestions:
              type: array
              items:
                type: string
    PlanSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [draft, active, archived]
        created_at:
          type: string
          format: date-time
        bed_count:
          type: integer
        total_area_sqft:
          type: number
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        request_id:
          type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "validation_error"
            message: "Invalid input parameters"
            details:
              field: "area.total_sqft"
              issue: "Must be between 10 and 50000"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "unauthorized"
            message: "Invalid or missing authentication token"
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "rate_limited"
            message: "API rate limit exceeded"
            details:
              retry_after: 60